requirements.txt
absl-py==2.1.0
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==23.1.0
backcall==0.2.0
batchbald-redux==2.0.5
bcrypt==4.2.0
beautifulsoup4==4.12.2
blackhc.project==2.7.0
bleach==4.1.0
cachetools==4.2.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
clarabel==0.7.1
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cryptography==43.0.0
cvxpy==1.5.2
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diffdist==0.1
dppy==0.3.2
ecos==2.0.14
entrypoints==0.4
exceptiongroup==1.2.0
executing==0.8.3
faiss-gpu==1.7.2
fastjsonschema==2.16.2
filelock==3.9.0
finch-clust==0.1.9
fonttools==4.50.0
fs==2.4.16
fs.sshfs==1.0.2
fsspec==2024.6.1
ftfy==6.2.3
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.64.1
gurobipy==11.0.3
huggingface-hub==0.24.2
idna==3.4
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.28.0
ipython==8.15.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
laaos==2.1.1
liac-arff==2.5.0
lightning-utilities==0.11.2
llvmlite==0.42.0
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
nets==0.0.3.1
networkx==3.2.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
openml==0.11.0
osqp==0.6.7.post0
overrides==7.4.0
packaging==23.1
pandas==2.2.2
pandocfilters==1.5.0
paramiko==3.4.0
parso==0.8.3
peppercorn==0.6
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.10.0
prefetch-generator==1.0.3
prometheus-client==0.14.1
prompt-toolkit==3.0.43
property-cached==1.6.4
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
Pygments==2.15.1
PyNaCl==1.5.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
qdldl==0.1.7.post4
referencing==0.30.2
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.10.1
scs==3.2.6
Send2Trash==1.8.2
setuptools==68.2.2
simclr==1.0.2
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
structlog==24.1.0
sympy==1.12
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
terminado==0.17.1
threadpoolctl==3.3.0
timm==1.0.7
tinycss2==1.2.1
toma==1.1.0
torchlars==0.1.2
torchmetrics==1.3.2
tornado==6.3.3
tqdm==4.66.2
traitlets==5.7.1
triton==2.2.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==3.0.3
wheel==0.41.2
wrapt==1.16.0
xmltodict==0.13.0
zipp==3.17.0
seaborn
